chore: move helper functions from context to utils
